Fixed error in DeckParser that caused it to skip lines. Removed contents of Deck
